Update 1-callouts.pillar
Re LibC's `time()`, the text says `(We leave it to the reader to look up the definition of the C function and determine the purpose of this argument.)` While the exact mechanism of the function is tangential, I think it's important and non-obvious why one can and should pass `NULL` as the argument to the spec. Indeed, declarations on some hosts do not specify it and the FFI callout seems to work in Pharo without it. If this is explained later in the booklet, we should mention that here; otherwise a more complete explanation would help